Refactored trait sections to allow for passing in ad-hoc details.
Extracted tags and tools to be their own components for rendering lists of content and plugged into the default slot for the relevant trait sections.

Simplified trait edit button management: now, the edit button doesn't show when the sheet is locked.

Extracted traits styles to a sheet component stylesheet that is imported into core scss.

Made special traits icon always visible, a star icon, and slightly larger.
