+++ src/sheets/actor/traits/Traits.svelte
@@ -1,20 +1,15 @@
 <script lang="ts">
   import { FoundryAdapter } from 'src/foundry/foundry-adapter';
-  import type {
-    CharacterSheetContext,
-    NpcSheetContext,
-    VehicleSheetContext,
-  } from 'src/types/types';
+  import type { ActorSheetContext } from 'src/types/types';
   import TraitSection from './TraitSection.svelte';
+  import TraitSectionTools from './TraitSectionTools.svelte';
   import { getContext } from 'svelte';
   import type { Readable } from 'svelte/store';
   import { settingStore } from 'src/settings/settings';
   import { error } from 'src/utils/logging';
+  import TraitSectionTags from './TraitSectionTags.svelte';
 
-  let context =
-    getContext<
-      Readable<CharacterSheetContext | VehicleSheetContext | NpcSheetContext>
-    >('context');
+  let context = getContext<Readable<ActorSheetContext>>('context');
   export let toggleable: boolean;
   export let useSenses: boolean = true;
   export let enableSpecialTraitsConfiguration: boolean = true;
@@ -32,6 +27,11 @@
     }
   }
 
+  // TODO: When Tidy takes over all data prep, give typings to tags
+  function getTags(obj: any): [key: string, value: string][] {
+    return Object.entries<string>(obj);
+  }
+
   const localize = FoundryAdapter.localize;
 </script>
 
@@ -40,103 +40,132 @@
   class:expanded={FoundryAdapter.tryGetFlag($context.actor, 'traitsExpanded')}
 >
   {#if useSenses && $context.senses}
+    {@const senses = getTags($context.senses)}
     <TraitSection
       title={localize('DND5E.Senses')}
       iconCssClass="fas fa-eye"
-      tags={Object.entries($context.senses)}
       configureButtonTitle={localize('DND5E.SensesConfig')}
       on:onConfigureClicked={() =>
         FoundryAdapter.renderActorSensesConfig($context.actor)}
-      {traitsExpanded}
-    />
+      show={traitsExpanded || !!senses.length}
+    >
+      <TraitSectionTags tags={senses} />
+    </TraitSection>
   {/if}
 
   {#if $context.traits?.traits?.languages}
+    {@const languages = getTags($context.traits?.traits?.languages?.selected)}
     <TraitSection
       traitCssClass={$context.traits?.traits?.languages?.cssClass ?? ''}
       title={localize('DND5E.Languages')}
       iconCssClass="fas fa-comment"
-      tags={Object.entries($context.traits?.traits?.languages?.selected)}
       configureButtonTitle={localize('DND5e.TraitConfig', {
         trait: localize('DND5E.Languages'),
       })}
       on:onConfigureClicked={() =>
         FoundryAdapter.renderTraitsSelector($context.actor, 'languages')}
-      {traitsExpanded}
-    />
+      show={traitsExpanded || !!languages.length}
+    >
+      <TraitSectionTags tags={languages} />
+    </TraitSection>
   {/if}
 
   {#if $context.traits.traits?.di}
+    {@const damageImmunities = getTags($context.traits.traits.di.selected)}
     <TraitSection
       traitCssClass={$context.traits.traits?.di?.cssClass ?? ''}
       title={localize('DND5E.DamImm')}
       iconCssClass="fas fa-heart"
-      tags={Object.entries($context.traits.traits.di.selected)}
       configureButtonTitle={localize('DND5E.TraitConfig', {
         trait: localize('DND5E.DamImm'),
       })}
       on:onConfigureClicked={() =>
         FoundryAdapter.renderTraitsSelector($context.actor, 'di')}
-      {traitsExpanded}
-    />
+      show={traitsExpanded || !!damageImmunities.length}
+    >
+      <TraitSectionTags tags={damageImmunities} />
+    </TraitSection>
   {/if}
 
   {#if $context.traits.traits?.dr}
+    {@const damageResistances = getTags($context.traits.traits.dr.selected)}
     <TraitSection
       traitCssClass={$context.traits.traits?.dr?.cssClass ?? ''}
       title={localize('DND5E.DamRes')}
       iconCssClass="far fa-heart"
-      tags={Object.entries($context.traits.traits.dr.selected)}
       configureButtonTitle={localize('DND5E.TraitConfig', {
         trait: localize('DND5E.DamRes'),
       })}
       on:onConfigureClicked={() =>
         FoundryAdapter.renderTraitsSelector($context.actor, 'dr')}
-      {traitsExpanded}
-    />
+      show={traitsExpanded || !!damageResistances.length}
+    >
+      <TraitSectionTags tags={damageResistances} />
+    </TraitSection>
   {/if}
 
   {#if $context.traits.traits?.dv}
+    {@const vulnerabilities = getTags($context.traits.traits.dv.selected)}
     <TraitSection
       traitCssClass={$context.traits.traits?.dv?.cssClass ?? ''}
       title={localize('DND5E.DamVuln')}
       iconCssClass="fas fa-heart-broken"
-      tags={Object.entries($context.traits.traits.dv.selected)}
       configureButtonTitle={localize('DND5E.TraitConfig', {
         trait: localize('DND5E.DamVuln'),
       })}
       on:onConfigureClicked={() =>
         FoundryAdapter.renderTraitsSelector($context.actor, 'dv')}
-      {traitsExpanded}
-    />
+      show={traitsExpanded || !!vulnerabilities.length}
+    >
+      <TraitSectionTags tags={vulnerabilities} />
+    </TraitSection>
+  {/if}
+
+  {#if $context.traits.traits?.dm}
+    {@const damageModifications = getTags($context.traits.traits.dm?.selected)}
+    <TraitSection
+      traitCssClass={$context.traits.traits?.dm?.cssClass ?? ''}
+      title={localize('DND5E.DamageModification.Label')}
+      iconCssClass="fas fa-heart-circle-plus"
+      configureButtonTitle={localize('DND5E.TraitConfig', {
+        trait: localize('DND5E.DamageModification.Label'),
+      })}
+      on:onConfigureClicked={() =>
+        FoundryAdapter.renderDamageModificationConfig($context.actor)}
+      show={traitsExpanded || !!damageModifications.length}
+    >
+      <!-- TODO: Add TraitSectionModifiers -->
+    </TraitSection>
   {/if}
 
   {#if $context.traits.traits?.ci}
+    {@const conditionImmunities = getTags($context.traits.traits.ci.selected)}
     <TraitSection
       traitCssClass={$context.traits.traits?.ci?.cssClass ?? ''}
       title={localize('DND5E.ConImm')}
       iconCssClass="fas fa-shield-virus"
-      tags={Object.entries($context.traits.traits.ci.selected)}
       configureButtonTitle={localize('DND5E.TraitConfig', {
         trait: localize('DND5E.ConImm'),
       })}
       on:onConfigureClicked={() =>
         FoundryAdapter.renderTraitsSelector($context.actor, 'ci')}
-      {traitsExpanded}
-    />
+      show={traitsExpanded || !!conditionImmunities.length}
+    >
+      <TraitSectionTags tags={conditionImmunities} />
+    </TraitSection>
   {/if}
 
   {#if $context.traits.traits?.weaponProf}
+    {@const weaponProfs = getTags($context.traits.traits.weaponProf.selected)}
     <TraitSection
       traitCssClass={$context.traits.traits?.weaponProf?.cssClass ?? ''}
       title={localize('DND5E.TraitWeaponProf')}
-      tags={Object.entries($context.traits.traits.weaponProf.selected)}
       configureButtonTitle={localize('DND5E.TraitConfig', {
         trait: localize('DND5E.TraitWeaponProf'),
       })}
       on:onConfigureClicked={() =>
         FoundryAdapter.renderTraitsSelector($context.actor, 'weapon')}
-      {traitsExpanded}
+      show={traitsExpanded || !!weaponProfs.length}
     >
       <svg
         slot="custom-icon"
@@ -147,27 +176,29 @@
       >
         <path
           d="M443.7,2.6c-2.2-2.5-5.8-3.3-8.8-1.8C417.9,9,358.8,36.1,304.4,43.5c-4.1,0.6-8-0.3-11.3-2.1c-1.5-0.8-2.4-2.5-2.4-4.2v0
-c0-10.7-8.7-19.4-19.5-19.4h-30.5c-10.7,0-19.5,8.7-19.5,19.4v0c0,1.7-0.9,3.4-2.4,4.2c-3.3,1.8-7.2,2.7-11.3,2.1
-C153.2,36.1,94.2,9,77.1,0.7c-3-1.5-6.6-0.7-8.8,1.8C56.3,16.4,21,64.9,21,154.9c0,83.4,30.3,143.6,44.3,166.9
-c2.9,4.9,10,4.7,12.8-0.2c10.5-19,34-59.1,62-91.7c0.4-0.5,0.9-1,1.3-1.5c28.3-30.7,79.9-9.8,79.9,32v232.2
-c0,10.7,8.7,19.4,19.5,19.4h30.5c10.7,0,19.5-8.7,19.5-19.4V260.4c0-41.8,51.5-62.7,79.9-32c0.4,0.5,0.9,1,1.3,1.5
-c28,32.6,51.5,72.7,62,91.7c2.8,5,9.9,5.1,12.8,0.2c14-23.3,44.3-83.4,44.3-166.9C491,64.9,455.7,16.4,443.7,2.6z"
+        c0-10.7-8.7-19.4-19.5-19.4h-30.5c-10.7,0-19.5,8.7-19.5,19.4v0c0,1.7-0.9,3.4-2.4,4.2c-3.3,1.8-7.2,2.7-11.3,2.1
+        C153.2,36.1,94.2,9,77.1,0.7c-3-1.5-6.6-0.7-8.8,1.8C56.3,16.4,21,64.9,21,154.9c0,83.4,30.3,143.6,44.3,166.9
+        c2.9,4.9,10,4.7,12.8-0.2c10.5-19,34-59.1,62-91.7c0.4-0.5,0.9-1,1.3-1.5c28.3-30.7,79.9-9.8,79.9,32v232.2
+        c0,10.7,8.7,19.4,19.5,19.4h30.5c10.7,0,19.5-8.7,19.5-19.4V260.4c0-41.8,51.5-62.7,79.9-32c0.4,0.5,0.9,1,1.3,1.5
+        c28,32.6,51.5,72.7,62,91.7c2.8,5,9.9,5.1,12.8,0.2c14-23.3,44.3-83.4,44.3-166.9C491,64.9,455.7,16.4,443.7,2.6z"
         />
       </svg>
+
+      <TraitSectionTags tags={weaponProfs} />
     </TraitSection>
   {/if}
 
   {#if $context.traits.traits?.armorProf}
+    {@const armorProfs = getTags($context.traits.traits.armorProf.selected)}
     <TraitSection
       traitCssClass={$context.traits.traits?.armorProf?.cssClass ?? ''}
       title={localize('DND5E.TraitArmorProf')}
-      tags={Object.entries($context.traits.traits.armorProf.selected)}
       configureButtonTitle={localize('DND5E.TraitConfig', {
         trait: localize('DND5E.TraitArmorProf'),
       })}
       on:onConfigureClicked={() =>
         FoundryAdapter.renderTraitsSelector($context.actor, 'armor')}
-      {traitsExpanded}
+      show={traitsExpanded || !!armorProfs.length}
     >
       <svg
         slot="custom-icon"
@@ -184,22 +215,28 @@ c28,32.6,51.5,72.7,62,91.7c2.8,5,9.9,5.1,12.8,0.2c14-23.3,44.3-83.4,44.3-166.9C4
       C474.9,362.7,471.7,291.8,471.7,198.8L471.7,198.8z"
         />
       </svg>
+
+      <TraitSectionTags tags={armorProfs} />
     </TraitSection>
   {/if}
 
   {#if $context.traits.tools}
+    {@const tools = Object.entries($context.tools)}
     <TraitSection
       traitCssClass={$context.traits.tools?.toolProf?.cssClass ?? ''}
       title={localize('DND5E.TraitToolProf')}
       iconCssClass="fas fa-hammer"
-      tools={Object.entries($context.tools)}
       configureButtonTitle={localize('DND5E.TraitConfig', {
         trait: localize('DND5E.TraitToolProf'),
       })}
       on:onConfigureClicked={() =>
         FoundryAdapter.renderToolSelector($context.actor)}
-      {traitsExpanded}
-    />
+      show={traitsExpanded || !!tools.length}
+    >
+      {#if tools.length}
+        <TraitSectionTools {tools} />
+      {/if}
+    </TraitSection>
   {/if}
 
   {#if $context.customActorTraits?.length}
@@ -224,7 +261,7 @@ c28,32.6,51.5,72.7,62,91.7c2.8,5,9.9,5.1,12.8,0.2c14-23.3,44.3-83.4,44.3-166.9C4
             );
           }
         }}
-        traitsExpanded={trait.alwaysShow || traitsExpanded}
+        show={trait.alwaysShow || traitsExpanded}
       />
     {/each}
   {/if}
@@ -243,7 +280,7 @@ c28,32.6,51.5,72.7,62,91.7c2.8,5,9.9,5.1,12.8,0.2c14-23.3,44.3-83.4,44.3-166.9C4
       {/if}
     </button>
   {/if}
-  {#if traitsExpanded && enableSpecialTraitsConfiguration && !$context.lockSensitiveFields}
+  {#if enableSpecialTraitsConfiguration && !$context.lockSensitiveFields}
     <button
       type="button"
       class="configure-special-traits inline-icon-button"
@@ -254,7 +291,7 @@ c28,32.6,51.5,72.7,62,91.7c2.8,5,9.9,5.1,12.8,0.2c14-23.3,44.3-83.4,44.3-166.9C4
         FoundryAdapter.renderActorSheetFlags($context.actor)}
       tabindex={$settingStore.useAccessibleKeyboardSupport ? 0 : -1}
     >
-      <i class="fas fa-cog" />
+      <i class="fas fa-star" />
     </button>
   {/if}
 </div>
